lambda functions/inference_lambda/requirements.txt
onnxruntime
boto3
numpy<2
Pillow